Refactor code out into a separate module
