+++ src/settings.ts
@@ -52,3 +52,5 @@ export const PROTECT_NVR_UNIFIOS_REFRESH_INTERVAL = 10;
 // How often, in seconds, should we heartbeat FFmpeg in two-way audio sessions. This should be less than 5 seconds, which is
 // FFmpeg's input timeout interval.
 export const PROTECT_TWOWAY_HEARTBEAT_INTERVAL = 3.5;
+
+export const PROTECT_DEFAULT_VIDEO_ENCODER: string = "libx264";
